create AboutNavLink component for specific styling
